django_blog/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.6
django-appconf==1.0.4
django-crispy-forms==1.11.2
django-formtools==2.3
django-haystack==3.0
django-imagekit==4.0.2
django-import-export==2.5.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-sslserver==0.22
djangorestframework==3.12.4
docopt==0.6.2
emoji==1.2.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.19.1
idna==2.10
jieba==0.42.1
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pilkit==2.0
Pillow==8.2.0
pipreqs==0.4.10
pycparser==2.20
Pygments==2.8.1
PyMySQL==1.0.2
pyparsing==2.4.7
pysolr==3.9.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.5
Whoosh==2.7.4
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
